+++ packages/types/src/create/registry.ts
@@ -22,10 +22,16 @@ import { Metadata } from '../metadata/Metadata';
 import { PortableRegistry } from '../metadata/PortableRegistry';
 import { lazyVariants } from './lazy';
 
+const DEFAULT_CALL_IDX = new Uint8Array(2);
+
 const l = logger('registry');
 
 function valueToString (v: { toString: () => string }): string {
   return v.toString();
+}
+
+function sortDecimalStrings (a: string, b: string): number {
+  return parseInt(a, 10) - parseInt(b, 10);
 }
 
 function getFieldArgs (lookup: PortableRegistry, fields: SiField[]): string[] {
@@ -133,6 +139,8 @@ function extractProperties (registry: TypeRegistry, metadata: Metadata): ChainPr
 
 export class TypeRegistry implements Registry {
   #classes = new Map<string, CodecClass>();
+
+  #defaultCallIndex: Uint8Array | null = null;
 
   #definitions = new Map<string, string>();
 
@@ -285,6 +293,12 @@ export class TypeRegistry implements Registry {
       this.#metadataCalls[`${section}`] && this.#metadataCalls[`${section}`][`${method}`],
       () => `findMetaCall: Unable to find Call with index [${section}, ${method}]/[${callIndex.toString()}]`
     );
+  }
+
+  // finds the default call for a specific runtime
+  // (since 0,0 may not be available)
+  public findMetaDefaultCall (): CallFunction {
+    return this.findMetaCall(this.#defaultCallIndex || DEFAULT_CALL_IDX);
   }
 
   // finds an error
@@ -503,6 +517,7 @@ export class TypeRegistry implements Registry {
   public setMetadata (metadata: Metadata, signedExtensions?: string[], userExtensions?: ExtDef): void {
     this.#metadata = metadata.asLatest;
     this.#metadataVersion = metadata.version;
+    this.#defaultCallIndex = null;
 
     // attach the lookup at this point (before injecting)
     this.setLookup(this.#metadata.lookup);
@@ -510,6 +525,22 @@ export class TypeRegistry implements Registry {
     injectExtrinsics(this, this.#metadata, this.#metadataVersion, this.#metadataCalls);
     injectErrors(this, this.#metadata, this.#metadataVersion, this.#metadataErrors);
     injectEvents(this, this.#metadata, this.#metadataVersion, this.#metadataEvents);
+
+    // set the default call index (the lowest section, the lowest method)
+    // in most chains this should be 0,0
+    const [defSection] = Object
+      .keys(this.#metadataCalls)
+      .sort(sortDecimalStrings);
+
+    if (defSection) {
+      const [defMethod] = Object
+        .keys(this.#metadataCalls[defSection])
+        .sort(sortDecimalStrings);
+
+      if (defMethod) {
+        this.#defaultCallIndex = new Uint8Array([parseInt(defSection, 10), parseInt(defMethod, 10)]);
+      }
+    }
 
     // setup the available extensions
     this.setSignedExtensions(